Refactor Rust parser: extract symbol source code logic to new module
- Moved the `get_symbol_source_code` function from `mod.rs` to a new `symbols.rs` file, improving code organization and separation of concerns.
- Updated imports in `mod.rs` to reflect the new module structure.
- Enhanced the symbol extraction logic by consolidating related functionality, making it easier to maintain and extend in the future.
- Added comprehensive tests for the new `get_symbol_source_code` function to ensure correctness and reliability of symbol extraction.
