+++ src/languages/rust/api/parsing/mod.rs
@@ -6,11 +6,13 @@ use tree_sitter::{Node, Parser};
 mod doc_comments;
 mod helpers;
 mod reexports;
+mod symbols;
 mod test_helpers;
 
-use doc_comments::{extract_inner_doc_comments, extract_outer_doc_comments};
-use helpers::{extract_attributes, extract_name, get_declaration_list, is_public};
+use doc_comments::extract_inner_doc_comments;
+use helpers::{extract_name, get_declaration_list, is_public};
 use reexports::extract_reexports;
+use symbols::get_symbol_source_code;
 
 pub fn parse_rust_file(content: &str, parser: &mut Parser) -> Result<RustFile, LaibraryError> {
     let tree = parser
@@ -79,65 +81,6 @@ fn extract_symbols_from_module(
     }
 
     Ok(symbols)
-}
-
-fn get_symbol_source_code(node: Node, source_code: &str) -> Result<String, LaibraryError> {
-    let mut source_code_with_docs = String::new();
-
-    if let Some(doc_comment) = extract_outer_doc_comments(&node, source_code)? {
-        source_code_with_docs.push_str(&doc_comment);
-    }
-
-    let attributes = extract_attributes(&node, source_code)?;
-    if !attributes.is_empty() {
-        let attributes_str = format!("{}\n", attributes.join("\n"));
-        source_code_with_docs.push_str(&attributes_str);
-    }
-
-    let symbol_source = match node.kind() {
-        "function_item" | "function_signature_item" => {
-            let mut cursor = node.walk();
-            let block_node = node
-                .children(&mut cursor)
-                .find(|n| n.kind() == "block")
-                .ok_or_else(|| LaibraryError::Parse("Failed to find function block".to_string()))?;
-            format!(
-                "{};",
-                &source_code[node.start_byte()..block_node.start_byte()].trim_end()
-            )
-        }
-        "trait_item" => {
-            let declaration_list = get_declaration_list(node).ok_or_else(|| {
-                LaibraryError::Parse("Failed to find trait declaration list".to_string())
-            })?;
-
-            let mut trait_source = String::new();
-            trait_source.push_str(&source_code[node.start_byte()..declaration_list.start_byte()]);
-            trait_source.push_str(" {\n");
-
-            let mut method_cursor = declaration_list.walk();
-            for method in declaration_list.children(&mut method_cursor) {
-                if method.kind() == "function_item" {
-                    let method_source = get_symbol_source_code(method, source_code)?;
-                    for line in method_source.lines() {
-                        trait_source.push_str("    ");
-                        trait_source.push_str(line);
-                        trait_source.push('\n');
-                    }
-                }
-            }
-
-            trait_source.push('}');
-            trait_source
-        }
-        _ => node
-            .utf8_text(source_code.as_bytes())
-            .map(|s| s.to_string())
-            .map_err(|e| LaibraryError::Parse(e.to_string()))?,
-    };
-
-    source_code_with_docs.push_str(&symbol_source);
-    Ok(source_code_with_docs)
 }
 
 #[cfg(test)]
